APIs-demo/pyproject.toml
[tool.poetry]
name = "tuts-for-se-thursday"
version = "0.1.0"
description = ""
authors = ["Bradleykibs <Bradleykibwana@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.29"
fastapi = "^0.110.1"
pydantic = {extras = ["email"], version = "^2.7.0"}
psycopg2 = "^2.9.9"
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
asyncpg = "^0.29.0"

[tool.poetry.group.dev.dependencies]
fastapi = "^0.110.1"
psycopg2-binary = "^2.9.9"
uvicorn = "^0.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
